调整配置文件读取逻辑

+++ Imitate-Soul-Knight-Project/Assets/Scripts/Data/Config/Level/WeaponConfigData.cs
@@ -6,21 +6,13 @@
 
 public class WeaponConfigData {
 
-	public int id;
-
-	public string name;
-
-	public string describe;
-
-	public string weaponUrl;
-
-	public float launchInterval;
-
-	public float damage;
-
-	public int launchCount;
-
-	public float bulletSpeed;
-
-	public string bulletUrl;
+	public int id; // 唯一标识
+	public string name; // 名称
+	public string describe; // 描述
+	public string weaponUrl; // 武器路径
+	public float launchInterval; // 发射间隔
+	public int damage; // 伤害
+	public int launchCount; // 发射数量
+	public int bulletSpeed; // 子弹速度
+	public string bulletUrl; // 子弹路径
 }
